flow/pyproject.toml
[project]
name = "flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elasticsearch>=9.2.1",
    "pyyaml>=6.0.3",
    "temporalio>=1.21.1",
    "tqdm>=4.67.1",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
